priceOracleApi/requirements.txt
pyyaml
pywaves
rsa
pycryptodome
psycopg2